Split and move holiday groups to a separate directory
