+++ holidays/groups/__init__.py
@@ -1,0 +1,17 @@
+#  python-holidays
+#  ---------------
+#  A fast, efficient Python library for generating country, province and state
+#  specific sets of holidays on the fly. It aims to make determining whether a
+#  specific date is a holiday as fast and flexible as possible.
+#
+#  Authors: dr-prodigy <dr.prodigy.github@gmail.com> (c) 2017-2023
+#           ryanss <ryanssdev@icloud.com> (c) 2014-2017
+#  Website: https://github.com/dr-prodigy/python-holidays
+#  License: MIT (see LICENSE file)
+
+from holidays.groups.buddhist import BuddhistCalendarHolidays
+from holidays.groups.chinese import ChineseCalendarHolidays
+from holidays.groups.christian import ChristianHolidays
+from holidays.groups.hindu import HinduCalendarHolidays
+from holidays.groups.international import InternationalHolidays
+from holidays.groups.islamic import IslamicHolidays
