+++ setup.cfg
@@ -43,9 +43,6 @@ current_version = 0.27
 extend-ignore = F821
 per-file-ignores =
     __init__.py:F401,F403
-    holidays/__init__.py:F401,F403
-    holidays/countries/__init__.py:F401
-    holidays/financial/__init__.py:F401
     scripts/generate_release_notes.py:T201
     tests/test_utils.py:T201
     tests/test_holiday_base.py:E203
@@ -57,7 +54,7 @@ strict = False
 [mypy-holidays.countries.*]
 disable_error_code = override
 
-[mypy-holidays.holiday_groups]
+[mypy-holidays.groups.*]
 disable_error_code = attr-defined
 
 [rstcheck]
